+++ ReleaseNotes.md
@@ -42,3 +42,4 @@
 ### vX.X.X (upcoming release)
 #### Features
 * Targets .NET Standard 1.3, .NET Standard 2.0 and .NET 4.5
+* Run migrations via SQL files